backend-python/requirements.txt
mysqlclient==2.0.3
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
APScheduler==3.10.4
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
dataclasses-json==0.5.14
doc2text==0.2.4
docopt==0.6.2
docx2txt==0.8
environs==9.5.0
flake8==6.1.0
Flask~=3.0.3
Flask-SSE==1.0.0
frozenlist==1.4.0
future==0.18.3
greenlet==2.0.2
grpcio==1.56.0
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.287
langsmith==0.0.36
MarkupSafe==2.1.3
marshmallow==3.20.1
mccabe==0.7.0
mime==0.1.0
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.6
numpy==1.25.2
openai==1.37.0
packaging==23.1
pandas==2.1.0
Pillow==10.0.0
protobuf==4.24.3
pycodestyle==2.11.0
pydantic==2.3.0
pydantic_core==2.6.3
pyflakes==3.1.0
PyJWT==2.8.0
elasticsearch==8.14.0
PyMySQL==1.1.0
pypdf==3.16.0
PyPDF2==3.0.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.0
requests==2.31.0
scipy==1.11.2
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.20
sseclient==0.0.27
stomp.py==8.1.0
tenacity==8.2.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
ujson==5.8.0
urllib3==2.0.4
websocket-client==1.6.3
Werkzeug~=3.0.0
yarl==1.9.2
celery==5.3.4
aliyun-python-sdk-core==2.13.3
zhipuai==1.0.7
qianfan==0.4.4
boto3==1.34.62
python-docx==1.1.0
cos-python-sdk-v5==1.9.26
dashscope==1.18.1
google-custom-search==2.0.9
volcengine-python-sdk==1.0.90
langdetect==1.0.9
pymilvus~=2.3.0
pyarrow==11.0.0
dataclasses~=0.6
genson==1.3.0
httpx==0.27.2
agora_python_server_sdk==2.1.4